Make action-navbar as a component
Since we have `Bucket` and `File` two different component that will
control which one should be display on action-navbar and call specified
method of service, so I update the action-navbar for control the UI.

* Add ActionNavbarService for control the type
* Move `bucket/action-navbar` to `layout/action-navbar`
* Add `onEnter` method to bucket state config for change type
