--- src/components/bucket/action-navbar/action-navbar.controller.js
@@ -1,12 +1,0 @@
-export default class ActionNavbarController {
-  /** @ngInject */
-  constructor($bucket) {
-    Object.assign(this, {
-      $bucket,
-    });
-  }
-
-  createBucket($event) {
-    this.$bucket.createDialog($event);
-  }
-}
