+++ src/components/bucket/bucket.js
@@ -5,26 +5,15 @@ import BucketController from './bucket.controller';
 import BucketService from './bucket.service';
 import BucketTemplate from './bucket.html';
 
-import ActionNavbarController from './action-navbar/action-navbar.controller';
-import ActionNavbarTemplate from './action-navbar/action-navbar.html';
-
 /** @ngInject */
 const route = $stateProvider => {
   $stateProvider.state('bucket', {
     url: '/bucket',
     parent: 'root',
-    views: {
-      '': {
-        controller: BucketController,
-        controllerAs: 'bucket',
-        template: BucketTemplate,
-      },
-      'action-navbar@root': {
-        template: ActionNavbarTemplate,
-        controller: ActionNavbarController,
-        controllerAs: 'nav',
-      },
-    },
+    controller: BucketController,
+    controllerAs: 'bucket',
+    template: BucketTemplate,
+    onEnter: $actionNav => $actionNav.setTypeToBucket(),
     ncyBreadcrumb: {
       label: 'All Buckets ( {{ bucket.data.length }} )',
     },
