+++ src/components/layout/action-navbar/action-navbar.service.js
@@ -1,0 +1,13 @@
+export default class ActionNavbarService {
+  constructor() {
+    this.type = 'Bucket';
+  }
+
+  setTypeToBucket() {
+    this.type = 'Bucket';
+  }
+
+  setTypeToFile() {
+    this.type = 'File';
+  }
+}
